requirements.txt
absl-py==1.2.0
addict==2.4.0
cachetools==5.2.0
certifi @ file:///C:/b/abs_4f5wo627a3/croots/recipe/certifi_1663615677642/work/certifi
charset-normalizer==2.1.1
colorama==0.4.5
cycler==0.11.0
Cython==0.29.32
easydict==1.10
filelock==3.8.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
huggingface-hub==0.10.0
idna==3.4
imageio==2.22.0
imgviz==1.5.1
importlib-metadata==4.12.0
joblib==1.2.0
kiwisolver==1.4.4
labelImg==1.8.6
labelme==5.0.2
lmdb==1.3.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.2.2
mmcv==1.6.2
natsort==8.2.0
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.1
opencv-python==4.4.0.46
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
prettytable==3.4.1
protobuf==3.19.6
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools-windows==2.0.0.2
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.3.0
PyYAML==6.0
QtPy==2.2.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.11
tokenizers==0.12.1
torch==1.7.1 
torchaudio==0.7.2  
torchtoolbox==0.1.8.2
torchvision==0.8.2 
tqdm==4.64.1
transformers==4.22.2
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wincertstore==0.2
yapf==0.32.0
zipp==3.8.1
